+++ docs/changelog.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 | Deprecated method                      | Substitution method                       | 
 |----------------------------------------|-------------------------------------------|
-| checkElements                          | getElementsTypes                           |
+| checkElements                          | getElementsTypes                          |
 | createElements                         | generateElements                          |
 | createNode                             | generateNode                              |
 | createLink                             | generateLink                              |
@@ -40,7 +40,62 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - New sc-event types: `AfterGenerateConnector`, `AfterGenerateEdge`, `BeforeEraseConnector`, `BeforeEraseEdge` were added.
 
+- All sc-types were redesigned to a common style. They were deprecated, new ones were added. See changes in the table below.
+  
+  | Deprecated                    | Substitution              |
+  |-------------------------------|---------------------------|
+  | ScType.EdgeUCommon            | ScType.CommonEdge         |
+  | ScType.EdgeDCommon            | ScType.CommonArc          |
+  | ScType.EdgeUCommonConst       | ScType.ConstCommonEdge    |
+  | ScType.EdgeDCommonConst       | ScType.ConstCommonArc     |
+  | ScType.EdgeAccess             | ScType.MembershipArc      |
+  | ScType.EdgeAccessConstPosPerm | ScType.ConstPermPosArc    |
+  | ScType.EdgeAccessConstNegPerm | ScType.ConstPermNegArc    |
+  | ScType.EdgeAccessConstFuzPerm | ScType.ConstFuzArc        |
+  | ScType.EdgeAccessConstPosTemp | ScType.ConstTempPosArc    |
+  | ScType.EdgeAccessConstNegTemp | ScType.ConstTempNegArc    |
+  | ScType.EdgeAccessConstFuzTemp | ScType.ConstFuzArc        |
+  | ScType.EdgeUCommonVar         | ScType.VarCommonEdge      |
+  | ScType.EdgeDCommonVar         | ScType.VarCommonArc       |
+  | ScType.EdgeAccessVarPosPerm   | ScType.VarPermPosArc      |
+  | ScType.EdgeAccessVarNegPerm   | ScType.VarPermNegArc      |
+  | ScType.EdgeAccessVarFuzPerm   | ScType.VarFuzArc          |
+  | ScType.EdgeAccessVarPosTemp   | ScType.VarTempPosArc      |
+  | ScType.EdgeAccessVarNegTemp   | ScType.VarTempNegArc      |
+  | ScType.EdgeAccessVarFuzTemp   | ScType.VarFuzArc          |
+  | ScType.NodeConst              | ScType.ConstNode          |
+  | ScType.NodeVar                | ScType.VarNode            |
+  | ScType.Link                   | ScType.NodeLink           |
+  | ScType.LinkClass              | ScType.NodeLinkClass      |
+  | ScType.NodeStruct             | ScType.NodeStructure      |
+  | ScType.LinkConst              | ScType.ConstNodeLink      |
+  | ScType.LinkConstClass         | ScType.ConstNodeLinkClass |
+  | ScType.NodeConstTuple         | ScType.ConstNodeTuple     |
+  | ScType.NodeConstStruct        | ScType.ConstNodeStructure |
+  | ScType.NodeConstRole          | ScType.ConstNodeRole      |
+  | ScType.NodeConstNoRole        | ScType.ConstNodeNonRole   |
+  | ScType.NodeConstClass         | ScType.ConstNodeClass     |
+  | ScType.NodeConstMaterial      | ScType.ConstNodeMaterial  |
+  | ScType.LinkVar                | ScType.VarNodeLink        |
+  | ScType.LinkVarClass           | ScType.VarNodeLinkClass   |
+  | ScType.NodeVarStruct          | ScType.VarNodeStructure   |
+  | ScType.NodeVarTuple           | ScType.VarNodeTuple       |
+  | ScType.NodeVarRole            | ScType.VarNodeRole        |
+  | ScType.NodeVarNoRole          | ScType.VarNodeNonRole     |
+  | ScType.NodeVarClass           | ScType.VarNodeClass       |
+  | ScType.NodeVarMaterial        | ScType.VarNodeMaterial    |
+
+- All sc-links are sc-nodes.
+- Types of actual and inactual temporal membership sc-arc were added.
+- All possible combinations of subtypes in sc-types have been added to the ScType.
+- Type `ScType.NodeAbstract` was removed.
+
 ### Added
+- ScType methods: `isConnector`, `isStructure`, `isActual`, `isInactual`, `isSuperclass`
+- All possible combinations of subtypes in sc-types into ScType
+- Type `ScType.NodeSuperclass`
+- Types: `ScType.Connector`, `ScType.Arc`
+- Types of actual and inactual temporal sc-arcs into ScType
 - sc-event types: `AfterGenerateConnector`, `AfterGenerateOutgoingArc`, `AfterGenerateIncomingArc`, `AfterGenerateEdge`, `BeforeEraseConnector`, `BeforeEraseOutgoingArc`, `BeforeEraseIncomingArc`, `BeforeEraseEdge`, `BeforeEraseElement`, `BeforeChangeLinkContent`
 - ScTemplate methods: `quintuple`
 - ScClient methods: `getElementsTypes`, `generateElements`, `generateNode`, `generateLink`, `generateConnector`, `generateElementsBySCs`, `eraseElements`, `searchLinksByContents`, `searchLinksByContentSubstrings`, `searchLinkContentsByContentSubstrings`, `searchByTemplate`, `generateByTemplate`, `createElementaryEventSubscriptions`, `destroyElementaryEventSubscriptions`, `searchKeynodes`
@@ -49,11 +104,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ScEvent` class was renamed to `ScEventSubscription`
 - `ScEventParams` class was renamed to `ScEventSubscriptionParams`
 
+### Fixed
+- Checking of all syntactic and semantic subtypes for types in ScType `merge` method.
+- Now sc-link is sc-node
+
 ### Deprecated
+- ScType methods: `isEdge`, `isStruct`
+- sc-types: `ScType.EdgeUCommon`, `ScType.EdgeDCommon`, `ScType.EdgeUCommonConst`, `ScType.EdgeDCommonConst`, `ScType.EdgeAccess`, `ScType.EdgeAccessConstPosPerm`, `ScType.EdgeAccessConstNegPerm`, `ScType.EdgeAccessConstFuzPerm`, `ScType.EdgeAccessConstPosTemp`, `ScType.EdgeAccessConstNegTemp`, `ScType.EdgeAccessConstFuzTemp`, `ScType.EdgeUCommonVar`, `ScType.EdgeDCommonVar`, `ScType.EdgeAccessVarPosPerm`, `ScType.EdgeAccessVarNegPerm`, `ScType.EdgeAccessVarFuzPerm`, `ScType.EdgeAccessVarPosTemp`, `ScType.EdgeAccessVarNegTemp`, `ScType.EdgeAccessVarFuzTemp`, `ScType.NodeConst`, `ScType.NodeVar`, `ScType.Link`, `ScType.LinkClass`, `ScType.NodeStruct`, `ScType.LinkConst`, `ScType.LinkConstClass`, `ScType.NodeConstTuple`, `ScType.NodeConstStruct`, `ScType.NodeConstRole`, `ScType.NodeConstNoRole`, `ScType.NodeConstClass`, `ScType.NodeConstMaterial`, `ScType.LinkVar`, `ScType.LinkVarClass`, `ScType.NodeVarStruct`, `ScType.NodeVarTuple`, `ScType.NodeVarRole`, `ScType.NodeVarNoRole`, `ScType.NodeVarClass`, `ScType.NodeVarMaterial`
 - ScTemplate methods: `tripleWithRelation`
 - ScClient methods: `checkElements`, `createElements`, `createNode`, `createLink`, `createEdge`, `createElementsBySCs`, `deleteElements`, `getLinksByContents`, `getLinksByContentSubstrings`, `getLinksContentsByContentSubstrings`, `templateSearch`, `templateGenerate`, `eventsCreate`, `eventsDestroy`, `findKeynodes`
 
 ### Removed
+- ScType method: `isAbstract`
+- Type `ScType.NodeAbstract`
 - sc-event types: `AddOutgoingEdge`, `AddIngoingEdge`, `RemoveOutgoingEdge`, `RemoveIngoingEdge`, `RemoveElement`, `ChangeContent`
 
 ## [v0.4.2]